Run tool-caml-tex tests in bytecode-only mode too
The tools/caml-tex used to be compiled only during opt.opt since this
slightly hackily ensured that the str and unix libraries were definitely
available. This target is not built in bytecode-only mode, so the tests
were marked as requiring the native compiler (even though caml-tex is a
bytecode tool!).

When tools/Makefile was merged, this quirk was corrected, so the tests
can now always be run.