Memprof.start replaces any existing profile in the domain, Memprof.is_sampling
The change to Memprof.start increases its compositionality while
conforming to the previous behaviour (it simply fails in fewer
situations). This is necessary for us to implement the Memprof
interface on top of the package memprof-limits.

The new function is_sampling is for clients that do want to fail
early, e.g. when detecting that two Memprof clients are interfering.