Pass variables explicitly to opam scripts
The ocaml-config package originally stored gen_ocaml_config.ml.in in the
files/ subdirectory, but policy changes in 2025 meaning it is now
downloaded from this repository. That removes the need for a separate
ocaml-config package, as its only purpose was to stop the script being
stored multiple times in the same repository.

Both gen_ocaml_config.ml.in and gen_ocaml-system_config.ml.in are
updated to have the opam variables they use passed explicitly on the
command line by the file using them. For the ocaml package, this means
that irrelevant options (for example spacetime) can be omitted and it
makes the scripts themselves somewhat easier to read.