Hashcons persistent identifier names
Maintain a table of strings used for persistent structures in the type
checker which is preserved across type-checker resets. This means that
other globals (in particular, the lazy thunks in Matching) can refer to
CamlinternalLazy and have the same physical string be used for that and
any references in the module being type-checked. Previously, whether
these thunks had been forced was physically observable in the marshalled
output of the module.