domain.c: terminate_backup_thread must synchronize with the backup thread
(see #14300)

`terminate_backup_thread` signals the `domain_cond` condition variable
while holding the domain lock, but it also broadcasts on the
interruptor's condition variable *without* holding its domain lock. If
the backup thread is in the state BT_IN_BLOCKING_SECTION, it waits on
the interruptor condition variable, and we want it to synchronize with
`terminate_backup_thread`, so that it does see the write to BT_TERMINATE.

To guarantee that the broadcast/wait is synchronizing, we tweak
`terminate_backup_thread` to temporarily take the interruptor's lock
when broadcasting on its condition variable.