[minor] fix a TSan report in shared_heap.c
Fixes #14334

In trunk `verify_heap` performs a `header_t*` dereference to compute
the header of large blocks, which can race against marking work done
in another domain. This race is benign (given that the `verify_heap`
code only checks that the block is not GARBAGE, and marking preserves
this property), and the standard way to ignore it is to use `Hd_hp`
which uses `volatile header_t *` dereference.