major_gc.c:ephe_mark: minor fix
The implementation in `trunk` has the main logic
`if (is_unmarked(key)) alive_data = 0`) placed in the `else` branch of
the `Tag_val(key) == Forward_tag` test, so it will not be checked on
non-shortcut `Forward_tag` value, which is a minor bug.

Change the code to first try to shortcut the Forward or Infix tag, and
then in any case check if the key is unmarked.

(noticed while reviewing #14336)